add request for user info - app have to do it when user open chat activity from notification

+++ app/src/main/kotlin/me/alexeyterekhov/vkfilter/GUI/ChooseUsersActivity/UserList/FriendListAdapter.kt
@@ -63,14 +63,14 @@ public class FriendListAdapter(
 
             name setText TextFormat.userTitle(user, false)
 
-            val id = user.id.toLong()
+            val id = user.id
             with (checkBox) {
                 setOnCheckedChangeListener(null)
                 setChecked(selected contains id)
                 setOnCheckedChangeListener {
                     view, checked ->
                     if (checked) {
-                        selected add id
+                        selected add id.toLong()
                     } else {
                         selected remove id
                     }
